+++ src/lib/server/migration-utils.ts
@@ -294,3 +294,35 @@ function populateTownHallLevel(thLevel: number) {
         spellCapacity: (spellHousing ?? 0) + (darkSpellHousing ?? 0),
     }
 }
+
+/**
+ * Updates TH data for new dark barracks level 11
+ * Adds data for new druid troop
+ */
+export async function v0_0_2_data_migrate(db: MySQL) {
+    // New dark barracks level 11 unlocked for town halls 14 and above
+    await db.query(`
+        UPDATE town_halls
+            SET maxDarkBarracks = 11
+        WHERE level >= 14
+    `);
+
+    // Insert unit data for new "druid" troop
+    const unitId = await db.insertOne('units', {
+        type: 'Troop',
+        name: 'Druid',
+        objectId: 40000123,
+        housingSpace: 16,
+        trainingTime: 150,
+        productionBuilding: 'Dark Elixir Barrack',
+        airTargets: true,
+        groundTargets: true
+    });
+    // Insert levels
+    await db.insertMany('unit_levels', [
+        { unitId, level: 1, barrackLevel: 11, laboratoryLevel: null },
+        { unitId, level: 2, barrackLevel: 11, laboratoryLevel: 12 },
+        { unitId, level: 3, barrackLevel: 11, laboratoryLevel: 13 },
+        { unitId, level: 4, barrackLevel: 11, laboratoryLevel: 14 },
+    ]);
+}